Přihlašování

+++ src/main/java/com/fandik/kino/controller/FilmController.java
@@ -1,9 +1,12 @@
 package com.fandik.kino.controller;
 
 import com.fandik.kino.entity.FilmEntity;
+import com.fandik.kino.entity.UzivatelEntity;
 import com.fandik.kino.service.FilmService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -34,14 +37,28 @@ public class FilmController {
     }
 
     @PostMapping
-    public ResponseEntity<FilmEntity> save(@RequestBody FilmEntity filmEntity) {
-        return new ResponseEntity<>(filmService.save(filmEntity), HttpStatus.CREATED);
+    public ResponseEntity<FilmEntity> save(@RequestBody FilmEntity filmEntity, @AuthenticationPrincipal UzivatelEntity uzivatelEntity) {
+        if (uzivatelEntity!=null) {
+            List<String> roles = uzivatelEntity.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toList());
+            if (roles.contains("ROLE_ADMIN")) {
+                return new ResponseEntity<>(filmService.save(filmEntity), HttpStatus.CREATED);
+            }
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteFilm(@PathVariable("id") Long id) {
-        filmService.deleteById(id);
-        return new ResponseEntity(HttpStatus.NO_CONTENT);
+    public ResponseEntity deleteFilm(@PathVariable("id") Long id, @AuthenticationPrincipal UzivatelEntity uzivatelEntity) {
+        if (uzivatelEntity!=null) {
+            List<String> roles = uzivatelEntity.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toList());
+            if (roles.contains("ROLE_ADMIN")) {
+                filmService.deleteById(id);
+                return new ResponseEntity(HttpStatus.NO_CONTENT);
+            }
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
 }
