+++ src/main/java/com/fandik/kino/controller/PredstaveniController.java
@@ -1,9 +1,12 @@
 package com.fandik.kino.controller;
 
 import com.fandik.kino.entity.PredstaveniEntity;
+import com.fandik.kino.entity.UzivatelEntity;
 import com.fandik.kino.service.PredstaveniService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -34,14 +37,28 @@ public class PredstaveniController {
     }
 
     @PostMapping
-    public ResponseEntity<PredstaveniEntity> save(@RequestBody PredstaveniEntity predstaveniEntity) {
-        return new ResponseEntity<>(predstaveniService.save(predstaveniEntity), HttpStatus.CREATED);
+    public ResponseEntity<PredstaveniEntity> save(@RequestBody PredstaveniEntity predstaveniEntity, @AuthenticationPrincipal UzivatelEntity uzivatelEntity) {
+        if (uzivatelEntity!=null) {
+            List<String> roles = uzivatelEntity.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toList());
+            if (roles.contains("ROLE_ADMIN")) {
+                return new ResponseEntity<>(predstaveniService.save(predstaveniEntity), HttpStatus.CREATED);
+            }
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity delete(@PathVariable("id") Long id) {
-        predstaveniService.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity delete(@PathVariable("id") Long id, @AuthenticationPrincipal UzivatelEntity uzivatelEntity) {
+        if (uzivatelEntity!=null) {
+            List<String> roles = uzivatelEntity.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toList());
+            if (roles.contains("ROLE_ADMIN")) {
+                predstaveniService.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
 }
